Supply right-hand-side callbacks for each equation.
NOTE that we need to substitute the dependencies for eq.rhs (implicitly pointed to within
eq.rhs somewhere) with the dependencies that come into the rhs callback. This currently
is /not/ done, as I need to fiddle with the UFL to do it, and so the replay won't actually work properly.