A first stab at annotating self-nonlinear equations.
Dolfin writes things in the form
F(u) = 0

libadjoint writes things in the form
A(u).u = b(u)

(it does this because it needs to write everything in
a big matrix to symbolically derive the adjoint equations).

So, when we get a solve(F = 0, u),
we actually annotate
M.u = M.u - F(u)

This requires some special modification of the RHS.__call__ method,
as in order to create the "source term", we first need to get
Dolfin to solve F(u) = 0, then put M.u on the RHS. That's
what the NonlinearRHS class does, instead.

I also added a debugging option "fussy_replay". When this is on,
the original forward run does things that are inefficient, like

solve(mass = action(mass, b), u)

instead of

u.assign(b)

but necessary for exact bit-for-bit compatibility with the forward replay
as annotated. I've set it defaulting to True for now, while the
whole setup is in development.

With these changes, the first nonlinear solve and assign is annotated
and replayed exactly. The later solves diverge a little
(hovering around 1e-10, 1e-11), because they currently start
the nonlinear solve from a zero initial guess, rather than the previous
value. (It's no coincidence that the discrepancy is approximately
the tolerance on the Newton solver.) This will be fixed in later commits.