A few minor fixes:
a) When adding two Vectors, if the x is None, don't add anything.
b) Make derivative_action robust to the case where the form doesn't
actually depend on the variable in question (used in initial guesses).
c) When passing an initial guess to a KrylovSolver, only pass an
initial guess if solving for a forward variable. (Forward initial guesses
tend to not be very good for adjoint solves.)
d) Make assign honour debugging["stop_annotating"].