Make the newton_constant.py test work.
The problem was the following. Imagine the user has done

dJdm = compute_gradient(J, ScalarParameter(m)).

This relies on the computation of each dF/dm, where F is the equation
to be solved. But if a nonlinear solve was registered, the 'F' returned
was just

Mu_new = Mu_true

where u_true was the solution of the nonlinear problem. In other words,
the form returned didn't say anything about the dependency on the parameter m.
So, the compute_gradient routine didn't work correctly.

The equation is still registered as Mu = Mu - F(u), but now the RHS also
has an attribute 'nonlinear_form' that contains the original form F, if
it be necessary for any particular purpose.