Add an option for the ReducedFunctionalNumPy to return its results in Euclidian space instead of the FunctionSpace used in the underlying discretisation.
This means that one can for example compute the norm of a derivative dj with:

numpy.dot(dj, dj)

instead of

inner(dj*dj*dx).

Similarly, one can compute the slope in a direction s with:

numpy.dot(dj, s).

The test euclidian_space checks the correctness of the implementation.