Implement Expressions with user-implemented derivative functions.
In many cases one has a set of ConstantControls which are used in an Expression
and which is then in turn used in the model. Up to now it was not possible
to differentiate such a model with respect to the Constant, because the
dependency of the Expression on the Constant was not detected by dolfin-adjoint.

This commit starts implementing support for such scenarios. The user must
specify the dependecy coefficients of the Expression and supply
the derivatives:

class SourceExpression(Expression):
    def __init__(self, c):
        self.c = c

    def eval(self, value, x):
        value[0] = float(self.c)**2

    def deval(self, value, x, coeff):
        value[0] = 2*self.c

    def dependencies(self):
        return [self.c]

    def copy(self):
        return SourceExpression(self.c)

Currently only first order derivatives are supported. A test lives in
expression_derivative/expression_derivative.py